dashboard

+++ back_end/src/main/java/com/swp_group4/back_end/entities/PaymentOrder.java
@@ -27,7 +27,7 @@ public class PaymentOrder {
     LocalDateTime paidDate;
     LocalDateTime dueDate;
     PaymentMethods paymentMethods;
-    Double total;
+    Long total;
     @Enumerated(EnumType.STRING)
     PaymentStatus status;
 }
