+++ back_end/src/main/java/com/swp_group4/back_end/responses/PaymentInfoResponse.java
@@ -18,6 +18,6 @@ public class PaymentInfoResponse {
     String paymentTitle;
     LocalDateTime paidDate;
     LocalDateTime dueDate;
-    Double price;
+    long price;
     PaymentStatus paymentStatus;
 }
